+++ Ladybird/Utilities.cpp
@@ -58,3 +58,16 @@ void platform_init()
     }();
 #endif
 }
+
+ErrorOr<Vector<String>> get_paths_for_helper_process(StringView process_name)
+{
+    auto application_path = TRY(ak_string_from_qstring(QCoreApplication::applicationDirPath()));
+    Vector<String> paths;
+
+    TRY(paths.try_append(TRY(String::formatted("./{}/{}", process_name, process_name))));
+    TRY(paths.try_append(TRY(String::formatted("{}/{}/{}", application_path, process_name, process_name))));
+    TRY(paths.try_append(TRY(String::formatted("{}/{}", application_path, process_name))));
+    TRY(paths.try_append(TRY(String::formatted("./{}", process_name))));
+    // NOTE: Add platform-specific paths here
+    return paths;
+}
