+++ Ladybird/WebDriver/CMakeLists.txt
@@ -5,7 +5,6 @@ set(SOURCES
     ${WEBDRIVER_SOURCE_DIR}/Session.cpp
     ${WEBDRIVER_SOURCE_DIR}/WebContentConnection.cpp
     ../Utilities.cpp
-    ../HelperProcess.cpp
     main.cpp
 )
