+++ Userland/Libraries/LibWebView/ViewImplementation.h
@@ -177,10 +177,6 @@ protected:
 
     virtual void create_client(EnableCallgrindProfiling = EnableCallgrindProfiling::No) { }
 
-#if !defined(AK_OS_SERENITY)
-    ErrorOr<NonnullRefPtr<WebView::WebContentClient>> launch_web_content_process(ReadonlySpan<String> candidate_web_content_paths, EnableCallgrindProfiling = EnableCallgrindProfiling::No, IsLayoutTestMode = IsLayoutTestMode::No, UseJavaScriptBytecode = UseJavaScriptBytecode::No);
-#endif
-
     void handle_web_content_process_crash();
 
     struct SharedBitmap {
